requirements.txt
playwright
python-dotenv
playwright-stealth
selenium
python-dotenv
